+++ fltk/src/macros/menu.rs
@@ -376,6 +376,30 @@ macro_rules! impl_menu_ext {
                 fn menu_frame(&self) -> $crate::enums::FrameType {
                     unsafe { $crate::enums::FrameType::from_i32([<$flname _menu_box>](self.inner.widget() as _)) }
                 }
+
+                fn mvalue(&self) -> Option<crate::menu::MenuItem> {
+                    unsafe {
+                        let ptr =
+                            [<$flname _mvalue>](self.inner.widget() as _) as *mut Fl_Menu_Item;
+                        if ptr.is_null() {
+                            None
+                        } else {
+                            Some(MenuItem::from_ptr(ptr))
+                        }
+                    }
+                }
+
+                fn prev_mvalue(&self) -> Option<crate::menu::MenuItem> {
+                    unsafe {
+                        let ptr =
+                            [<$flname _prev_mvalue>](self.inner.widget() as _) as *mut Fl_Menu_Item;
+                        if ptr.is_null() {
+                            None
+                        } else {
+                            Some(MenuItem::from_ptr(ptr))
+                        }
+                    }
+                }
             }
         }
     };