+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # [1.4.33] - TBA
 - Add Window::set_on_top() support on windows by @caesay.
+- Add MenuExt::mvalue() and prev_mvalue().
 - Update docs and improve editor examples by @ntrrgc.
 - Refactor Window::raw_handle() code.
 - image::Pixmap improvements.